Refactor: Remove redundant isLikelyKeywordStart method and consolidate optimization
- Remove isLikelyKeywordStart() method as its logic is duplicated by the character-based switch statement
- The switch statement already handles all keyword starting characters (a,d,e,f,g,h,i,m,o,s,u,w)
- Non-keyword characters naturally fall through to handleWord() call
- Eliminate code duplication while maintaining performance optimizations
- All 60 SqlTokenizer tests pass with 100% success rate

This cleanup simplifies the codebase without affecting functionality or performance.
