+++ doma-core/src/jmh/java/org/seasar/doma/internal/util/SqlTokenUtilBenchmark.java
@@ -26,8 +26,8 @@ import org.openjdk.jmh.annotations.State;
 import org.openjdk.jmh.infra.Blackhole;
 
 /**
- * JMH benchmark for comparing performance between FastCharClassifier, the optimized SqlTokenUtil,
- * and the original ClassicSqlTokenUtil.
+ * JMH benchmark for comparing performance between the optimized SqlTokenUtil and the original
+ * ClassicSqlTokenUtil.
  */
 @SuppressWarnings("deprecation")
 @BenchmarkMode(Mode.AverageTime)
@@ -62,13 +62,6 @@ public class SqlTokenUtilBenchmark {
   }
 
   @Benchmark
-  public void fastCharClassifierIsWordPart(Blackhole bh) {
-    for (char c : testChars) {
-      bh.consume(FastCharClassifier.isWordPart(c));
-    }
-  }
-
-  @Benchmark
   public void optimizedIsWordPart(Blackhole bh) {
     for (char c : testChars) {
       bh.consume(SqlTokenUtil.isWordPart(c));
@@ -79,13 +72,6 @@ public class SqlTokenUtilBenchmark {
   public void classicIsWordPart(Blackhole bh) {
     for (char c : testChars) {
       bh.consume(ClassicSqlTokenUtil.isWordPart(c));
-    }
-  }
-
-  @Benchmark
-  public void fastCharClassifierIsWhitespace(Blackhole bh) {
-    for (char c : testChars) {
-      bh.consume(FastCharClassifier.isWhitespace(c));
     }
   }
 
@@ -104,19 +90,8 @@ public class SqlTokenUtilBenchmark {
   }
 
   @Benchmark
-  public void fastCharClassifierCombined(Blackhole bh) {
-    // Test combined usage pattern with FastCharClassifier direct calls
-    for (char c : testChars) {
-      boolean isWord = FastCharClassifier.isWordPart(c);
-      boolean isSpace = FastCharClassifier.isWhitespace(c);
-      bh.consume(isWord);
-      bh.consume(isSpace);
-    }
-  }
-
-  @Benchmark
   public void optimizedCombined(Blackhole bh) {
-    // Test combined usage pattern
+    // Test combined usage pattern with optimized SqlTokenUtil
     for (char c : testChars) {
       boolean isWord = SqlTokenUtil.isWordPart(c);
       boolean isSpace = SqlTokenUtil.isWhitespace(c);
@@ -127,7 +102,7 @@ public class SqlTokenUtilBenchmark {
 
   @Benchmark
   public void classicCombined(Blackhole bh) {
-    // Test combined usage pattern
+    // Test combined usage pattern with original ClassicSqlTokenUtil
     for (char c : testChars) {
       boolean isWord = ClassicSqlTokenUtil.isWordPart(c);
       boolean isSpace = ClassicSqlTokenUtil.isWhitespace(c);
